Using $wgResourceLoaderLESSVars for most LESS vars
This change should make it easier to adapt the look of BlueSpiceSkin to
the requirements of customers. It also enables extensions like FlexiSkin
to modify the variables by using 'ResourceLoaderGetLessVars' (1.25+) hook.

I don't know why, but having 'bs-skin-path' variable in the array breaks
the function. This should be subject to future improvements.

Change-Id: I6ad1a5719acb1a78b20be7fb602f5112524c373d